0.0.9: Another bugfix

+++ ProkaryoteAssembly/prokaryote_assemble.py
@@ -62,7 +62,7 @@ def assemble(fwd_reads, rev_reads, out_dir, cleanup):
 def total_cleanup(input_dir: Path):
     all_files = list(input_dir.glob("*"))
     for f in all_files:
-        if f.suffix is not ".fasta":
+        if f.suffix is not ".fasta" and f.is_file():
             os.remove(str(f))
     try:
         shutil.rmtree(str(input_dir / 'pilon'))
